+++ apps/web/src/app/(main)/(docs)/components/tab/page.tsx
@@ -1,3 +1,4 @@
+import { Code } from "@/components/code";
 import { VariantPage } from "@/components/docs/variant-page";
 import { VariantWrapper } from "@/components/docs/variant-wrapper";
 import { type Metadata } from "next";
@@ -12,9 +13,15 @@ import { TabOutside } from "./variants/outside";
 import { TabRadius } from "./variants/radius";
 import { TabSizes } from "./variants/sizes";
 import { TabVariants } from "./variants/variants";
+import { TabWithIcons } from "./variants/with-icons";
+
+const title = "Tab";
+const description =
+  "Tabs structure content into organized sections, providing users with a streamlined way to navigate between them.";
 
 export const metadata: Metadata = {
-  title: "Tab",
+  title,
+  description,
 };
 
 const code = <T extends VariantTypes["tab"][number]>(variant: T) =>
@@ -22,14 +29,46 @@ const code = <T extends VariantTypes["tab"][number]>(variant: T) =>
 
 const Tab = () => {
   return (
-    <VariantPage heading="Tab">
+    <VariantPage heading="Alert" description={description}>
       <VariantWrapper heading="Default" code={code("default")}>
         <TabDefault />
       </VariantWrapper>
-      <VariantWrapper heading="Disabled" code={code("disabled")}>
+
+      <VariantWrapper
+        heading="Variants"
+        code={code("variants")}
+        description={
+          <div>
+            Use the <Code>variant</Code> prop to change the visual style of the
+            tabs. Values can be <Code>underlined</Code>,<Code>bordered</Code>,
+            <Code>solid</Code> or <Code>light</Code>.
+          </div>
+        }
+      >
+        <TabVariants />
+      </VariantWrapper>
+
+      <VariantWrapper
+        heading="Disabled"
+        code={code("disabled")}
+        description={
+          <div>
+            Use the <Code>isDisabled</Code> prop to to disable a tab components.
+          </div>
+        }
+      >
         <TabDisabled />
       </VariantWrapper>
-      <VariantWrapper heading="Disabled Item" code={code("disabled-item")}>
+
+      <VariantWrapper
+        heading="Disabled Item"
+        code={code("disabled-item")}
+        description={
+          <div>
+            Use the <Code>isDisabled</Code> prop to to disable a tab.
+          </div>
+        }
+      >
         <TabDisabledItem />
       </VariantWrapper>
       <VariantWrapper heading="Sizes" code={code("sizes")}>
@@ -41,10 +80,22 @@ const Tab = () => {
       <VariantWrapper heading="Colors" code={code("colors")}>
         <TabColors />
       </VariantWrapper>
-      <VariantWrapper heading="Variants" code={code("variants")}>
-        <TabVariants />
+
+      <VariantWrapper
+        heading="With Icons"
+        code={code("with-icons")}
+        // description={
+        //   <div>
+        //     Use the <Code>variant</Code> prop to change the visual style of the
+        //     tabs. Values can be <Code>underlined</Code>,<Code>bordered</Code>,
+        //     <Code>solid</Code> or <Code>light</Code>.
+        //   </div>
+        // }
+      >
+        <TabWithIcons />
       </VariantWrapper>
-      <VariantWrapper heading="As" code={code("as")}>
+
+      <VariantWrapper heading="As Link" code={code("as")}>
         <TabAs />
       </VariantWrapper>
       <VariantWrapper heading="Outside" code={code("outside")}>