+++ apps/web/src/app/(main)/(docs)/components/tab/variants/disabled.tsx
@@ -1,28 +1,26 @@
-import { Tab, Tabs } from "@jamsr-ui/react";
-import { TabContent } from "../shared";
+import { Tab, Tabs, Card, CardContent } from "@jamsr-ui/react";
+
+const TabContent = () => (
+  <Card>
+    <CardContent>
+      Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
+      tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim
+      veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea
+      commodo consequat.
+    </CardContent>
+  </Card>
+);
 
 export const TabDisabled = () => {
   return (
-    <Tabs
-      defaultValue="photos"
-      isDisabled
-    >
-      <Tab
-        value="photos"
-        heading="Photos"
-      >
+    <Tabs defaultValue="photos" isDisabled variant="bordered">
+      <Tab value="photos" heading="Photos">
         <TabContent />
       </Tab>
-      <Tab
-        value="music"
-        heading="Music"
-      >
+      <Tab value="music" heading="Music">
         <TabContent />
       </Tab>
-      <Tab
-        value="videos"
-        heading="Videos"
-      >
+      <Tab value="videos" heading="Videos">
         <TabContent />
       </Tab>
     </Tabs>