+++ apps/web/src/app/providers/app.tsx
@@ -14,6 +14,7 @@ export const AppProvider = (props: Props) => {
       <Analytics />
       <ToastProvider />
       <UIStylesProvider
+      
       // button={{
       //   className: "",
       //   color: "danger",