+++ packages/components/dialog/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jamsr-ui/dialog
 
+## 0.0.74
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/shared-icons@0.0.56
+  - @jamsr-ui/button@0.0.66
+
 ## 0.0.73
 
 ### Patch Changes